Extended with a logic to check if any team has an empty value

+++ LiveFootballScoreBoard/Services/IFootballScoreBoardService.cs
@@ -4,9 +4,9 @@ namespace LiveFootballScoreBoard.Services
 {
 	public interface IFootballScoreBoardService
 	{
-		ExecutionResult<int> StartFootballMatch(string homeTeam, string awayTeam);
-		ExecutionResult UpdateMatchScore(int matchId, ushort homeTeamScore, ushort awayTeamScore);
-		ExecutionResult FinishMatch(int matchId);
+		ExecutionResult<long> StartFootballMatch(string homeTeam, string awayTeam);
+		ExecutionResult UpdateMatchScore(long matchId, ushort homeTeamScore, ushort awayTeamScore);
+		ExecutionResult FinishMatch(long matchId);
 		ExecutionResult<IList<FootballMatch>> GetMatchesScoreSummary();
 	}
 }
